Finer grained control of what to do to servers

+++ start-slaves.sh
@@ -7,6 +7,26 @@ echo ========================================================
 echo Start slave script
 echo ========================================================
 
+for var in "$@"
+do
+    if [[ "$var" == "skip.copy" ]] ; then
+       SKIP_COPY=1
+    elif [[ "$var" == "force.copy" ]] ; then
+       FORCE_COPY=1
+    elif [[ "$var" == "start" ]] ; then
+       START=1
+    elif [[ "$var" == "stop" ]] ; then
+       STOP=1
+    else
+        echo $var not supported, supported options: skip.copy,force.copy
+    fi
+done
+
+if [[ $STOP != "1" ]] && [[ $START != "1" ]] ; then
+    START=1
+    STOP=1
+fi
+
 IFS=$'\n'
 dc_addr=
 #Read the DC
@@ -49,36 +69,46 @@ do
 	  echo ----------------------------------------
 	  echo Contacting $slave_host $slave_addr
 
-          #Get the slave's revision
-          rm -f work/temp-rev.txt
-	  scp ec2-user@$slave_addr:~/slave/current-rev.txt work/temp-rev.txt
-	  slaverev=`cat work/temp-rev.txt`
-	  echo slave rev: $slaverev
+	  if [ "$SKIP_COPY" != "1" ] ; then
 
-	  if [[ $slaverev != $myrev ]] ; then
-	      #Slave revision was different - copy files to slave
-	      echo copying files to slave $slave_addr
-	      #Remove the existing files in the slave folder
-	      ssh $slave_addr rm -rf slave/*
-	      #Copy the current rev to the slave
-	      scp ../jboss-as/build/target/current-rev.txt ec2-user@$slave_addr:~/slave/current-rev.txt
-	      #Copy the zipped AS installation to the slave
-	      scp ../jboss-as/build/target/jboss-as.zip $slave_addr:~/slave/jboss-as.zip
-	      #Unzip the zipped AS installation
-	      echo unzipping as installation on $slave_addr
-	      ssh $slave_addr 'unzip -q ~/slave/jboss-as.zip -d ~/slave'
-	  fi
+              #Get the slave's revision
+              rm -f work/temp-rev.txt
+	      scp ec2-user@$slave_addr:~/slave/current-rev.txt work/temp-rev.txt
+	      slaverev=`cat work/temp-rev.txt`
+	      echo slave rev: $slaverev
 
-          #Overwrite the host name in host-slave.xml with the name of this slave from $slave_host (i.e. from the Name tag)
-	  rm -f work/temp-host-slave.xml
-	  cp config/host-slave.xml work/temp-host-slave.xml
-	  perl -pi -e 's/__HOST_NAME__/'$slave_host'/g' work/temp-host-slave.xml
-	  #ssh -o "StrictHostKeyChecking no" $slave_addr 'rm ~/'$BUILT_JBOSS'/domain/configuration/host-slave.xml'
-	  scp work/temp-host-slave.xml $slave_addr:~/slave/$BUILT_JBOSS/domain/configuration/host-slave.xml
+	      if [[ $slaverev != $myrev ]] || [[ $FORCE_COPY == "1" ]] ; then
+	          #Slave revision was different - copy files to slave
+	          echo copying files to slave $slave_addr
+	          #Remove the existing files in the slave folder
+	          ssh $slave_addr rm -rf slave/*
+	          #Copy the current rev to the slave
+	          scp ../jboss-as/build/target/current-rev.txt ec2-user@$slave_addr:~/slave/current-rev.txt
+	          #Copy the zipped AS installation to the slave
+	          scp ../jboss-as/build/target/jboss-as.zip $slave_addr:~/slave/jboss-as.zip
+	          #Unzip the zipped AS installation
+	          echo unzipping as installation on $slave_addr
+	          ssh $slave_addr 'unzip -q ~/slave/jboss-as.zip -d ~/slave'
+	      fi
 
-          echo start slave
+              #Overwrite the host name in host-slave.xml with the name of this slave from $slave_host (i.e. from the Name tag)
+	      rm -f work/temp-host-slave.xml
+	      cp config/host-slave.xml work/temp-host-slave.xml
+	      perl -pi -e 's/__HOST_NAME__/'$slave_host'/g' work/temp-host-slave.xml
+	      #ssh -o "StrictHostKeyChecking no" $slave_addr 'rm ~/'$BUILT_JBOSS'/domain/configuration/host-slave.xml'
+	      scp work/temp-host-slave.xml $slave_addr:~/slave/$BUILT_JBOSS/domain/configuration/host-slave.xml
+	  fi 
+
 	  #Go to the slave's bin directory, kill all running java processes and start the domain in the background
-          ssh  -o "StrictHostKeyChecking no" $slave_addr 'cd ~/slave/'$BUILT_JBOSS'/bin; killall -9 java ; nohup ./domain.sh --host-config=host-slave.xml -Djboss.bind.address.management=0.0.0.0 -Djboss.bind.address=0.0.0.0 -Djboss.domain.master.address='$dc_addr' < /dev/null > /dev/null 2>/dev/null &'
+
+	  if [[ $STOP == "1" ]] ; then 
+              echo stop slave
+              ssh  -o "StrictHostKeyChecking no" $slave_addr 'killall -9 java'
+	  fi
+	  if [[ $START == "1" ]] ; then 
+              echo start slave
+              ssh  -o "StrictHostKeyChecking no" $slave_addr 'cd ~/slave/'$BUILT_JBOSS'/bin; nohup ./domain.sh --host-config=host-slave.xml -Djboss.bind.address.management=0.0.0.0 -Djboss.bind.address=0.0.0.0 -Djboss.domain.master.address='$dc_addr' < /dev/null > /dev/null 2>/dev/null &'
+	  fi
 
           #No idea if this makes any difference, but clear these variables
 	  slave_host=
